scripts/requirements.txt
boto3
python-dotenv
httpx
cryptography